Set the playlist context menu popup location relative to the mouse cursor to ensure a better layout on the user's screen.
There doesn't appear to be a way to get the laid out X / Y of a node in JavaFX
so anchoring relative to the mouse seems better - but not infallible.

E.g. if the mouse is at ScreenX, ScreenY (0,0) and the user selects the button
by keyboard, then the menu will be back to displaying off the stage again.